requirements.txt
python-telegram-bot==20.7
ffmpeg-python==0.2.0
python-dotenv==1.0.0
pytube==15.0.0 
